+++ geode-core/src/distributedTest/java/org/apache/geode/cache/query/dunit/QueryUsingPoolDUnitTest.java
@@ -558,7 +558,7 @@ public class QueryUsingPoolDUnitTest extends JUnit4CacheTestCase {
       assertEquals(1, results.size());
       Object result = results.iterator().next();
       assertTrue(result instanceof Integer);
-      int resultInt = ((Integer) result).intValue();
+      int resultInt = (Integer) result;
       assertEquals(resultInt, 100);
 
       // query with leading/trailing spaces
@@ -1865,8 +1865,8 @@ public class QueryUsingPoolDUnitTest extends JUnit4CacheTestCase {
 
     @Override
     public int compare(Object obj1, Object obj2) {
-      int obj1Id = ((Integer) obj1).intValue();
-      int obj2Id = ((Integer) obj2).intValue();
+      int obj1Id = (Integer) obj1;
+      int obj2Id = (Integer) obj2;
       if (obj1Id > obj2Id) {
         return 1;
       } else if (obj1Id < obj2Id) {
@@ -1881,8 +1881,8 @@ public class QueryUsingPoolDUnitTest extends JUnit4CacheTestCase {
 
     @Override
     public int compare(Object obj1, Object obj2) {
-      int obj1Id = ((Integer) ((Struct) obj1).get("id")).intValue();
-      int obj2Id = ((Integer) ((Struct) obj2).get("id")).intValue();
+      int obj1Id = (Integer) ((Struct) obj1).get("id");
+      int obj2Id = (Integer) ((Struct) obj2).get("id");
       if (obj1Id > obj2Id) {
         return 1;
       } else if (obj1Id < obj2Id) {
