+++ geode-core/src/distributedTest/java/org/apache/geode/distributed/internal/streaming/StreamingOperationOneDUnitTest.java
@@ -140,7 +140,7 @@ public class StreamingOperationOneDUnitTest extends JUnit4DistributedTestCase {
       // sort the input streams
       for (final Object o : chunkMap.entrySet()) {
         Map.Entry entry = (Map.Entry) o;
-        int seqNum = ((Integer) entry.getKey()).intValue();
+        int seqNum = (Integer) entry.getKey();
         objList = (List) entry.getValue();
         arrayOfLists[seqNum] = objList;
       }
@@ -151,7 +151,7 @@ public class StreamingOperationOneDUnitTest extends JUnit4DistributedTestCase {
         Integer nextInteger;
         while (itr.hasNext()) {
           nextInteger = (Integer) itr.next();
-          if (nextInteger.intValue() != expectedInt) {
+          if (nextInteger != expectedInt) {
             logger.severe("nextInteger.intValue() != expectedInt");
             return;
           }
